修改外部协议
